pyproject.toml
[tool.poetry]
name = "whiteboard-api"
version = "0.1.0"
description = ""
authors = ["MasonChance <mason.fryberger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.2"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"
postgres = "^3.0.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
